@@ -7,7 +7,6 @@
     "test": "jest --coverage --env=node",
     "snyk-test": "snyk test",
     "lint": "eslint ./src --ext .ts --output-file ./eslint-report.json --format json",
-    "preinstall": "npx npm-force-resolutions",
     "postbuild": "bash package.sh",
     "schematic:list": "npm run build && schematics ./dist/schematics/collection.json:list employee --path=\"./dist/\" --openApiJsonUrl=https://im-wa-empo-nrcrn.azurewebsites.net/swagger/v1/swagger.json --dry-run=false --force --verbose --appPrefix=nrcrn",
     "schematic:module:bld": "npm run build && schematics ./dist/schematics/collection.json:module building --path=\"./dist/\" --openApiJsonUrl=https://im-wa-unto-nrcrn.azurewebsites.net/swagger/v1/swagger.json --dry-run=false --force --verbose --appPrefix=nrcrn",
@@ -28,34 +27,32 @@
   "license": "MIT",
   "schematics": "./src/collection.json",
   "dependencies": {
-    "@angular-devkit/core": "^14.1.2",
-    "@angular-devkit/schematics": "^14.1.2",
-    "axios": "^0.27.2",
-    "find-up": "^5.0.0",
+    "@angular-devkit/core": "16.0.2",
+    "@angular-devkit/schematics": "16.0.2",
+    "axios": "^1.4.0",
     "lodash": "^4.17.21",
     "pluralize": "^8.0.0",
-    "rxjs": "^6.6.7"
+    "rxjs": "^7.8.1"
   },
   "devDependencies": {
-    "@angular-devkit/schematics-cli": "^14.1.2",
-    "@angular/compiler": "^14.1.2",
-    "@angular/core": "^14.1.2",
-    "@nrwl/angular": "^14.5.6",
-    "@nrwl/eslint-plugin-nx": "^14.5.6",
-    "@types/jest": "^27.4.1",
-    "@types/lodash": "^4.14.182",
-    "@types/node": "^16.11.48",
+    "@angular-devkit/schematics-cli": "16.0.2",
+    "@angular/compiler": "16.0.2",
+    "@angular/core": "16.0.2",
+    "@nx/angular": "16.2.1 ",
+    "@nx/eslint-plugin": "16.2.1",
+    "@types/jest": "^29.5.1",
+    "@types/lodash": "^4.14.194",
+    "@types/node": "^18.16.13",
     "@types/pluralize": "0.0.29",
-    "@typescript-eslint/eslint-plugin": "^5.33.0",
-    "@typescript-eslint/parser": "^5.33.0",
-    "eslint": "^8.22.0",
-    "eslint-config-prettier": "8.5",
-    "jest": "^27.5.1",
-    "jest-junit": "^13.2.0",
-    "npm-force-resolutions": "^0.0.10",
+    "@typescript-eslint/eslint-plugin": "^5.59.6",
+    "@typescript-eslint/parser": "^5.59.6",
+    "eslint": "^8.40.0",
+    "eslint-config-prettier": "8.8",
+    "jest": "^29.5.0",
+    "jest-junit": "^16.0.0",
     "schematics": "^0.2.0",
-    "snyk": "^1.1064.0",
-    "ts-jest": "^27.1.4",
-    "typescript": "^4.4.4"
+    "snyk": "^1.1159.0",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.0.4"
   }
 }
