@@ -1,32 +1,32 @@
-import { Tree } from '@angular-devkit/schematics';
+import { Tree } from "@angular-devkit/schematics";
 import {
   SchematicTestRunner,
   UnitTestTree,
-} from '@angular-devkit/schematics/testing';
-import * as path from 'path';
-import { IOptions } from '../shared';
-import { readFirst } from '@nrwl/angular/testing/src/testing-utils';
-import * as pluralize from 'pluralize';
-import { classify } from '@angular-devkit/core/src/utils/strings';
+} from "@angular-devkit/schematics/testing";
+import * as path from "path";
+import { IOptions } from "../shared";
+import * as pluralize from "pluralize";
+import { classify } from "@angular-devkit/core/src/utils/strings";
 
-const collectionPath = path.join(__dirname, '../collection.json');
+const collectionPath = path.join(__dirname, "../collection.json");
 
-describe('imng-list', () => {
-  test('generation works', async () => {
-    const runner = new SchematicTestRunner('schematics', collectionPath);
+describe("imng-list", () => {
+  test("generation works", async () => {
+    const runner = new SchematicTestRunner("schematics", collectionPath);
     const options: IOptions = {
-      name: 'certification',
+      name: "certification",
       openApiJsonUrl:
-        'https://raw.githubusercontent.com/ikemtz/AngularMonoRepo/master/schematics/open-api-docs/nrcrn-cert-odata.json',
-      path: './test',
+        "https://raw.githubusercontent.com/ikemtz/AngularMonoRepo/master/schematics/open-api-docs/nrcrn-cert-odata.json",
+      path: "./test",
       swaggerProperties: [],
-      storeName: 'certifications',
-      appPrefix: 'nrcrn',
+      storeName: "certifications",
+      appPrefix: "nrcrn",
     };
-    const tree: UnitTestTree = await readFirst(
-      runner.runSchematicAsync('imng-list', options, Tree.empty())
+    const tree: UnitTestTree = await runner.runSchematic(
+      "imng-list",
+      options,
+      Tree.empty()
     );
-
     expect(tree.files).toEqual([
       `/test/${pluralize(options.name)}-list/index.ts`,
       `/test/${pluralize(options.name)}-list/list.component.html`,
