@@ -4,8 +4,7 @@ import {
   UnitTestTree,
 } from '@angular-devkit/schematics/testing';
 import * as path from 'path';
-import { IOptions } from '../shared';
-import { readFirst } from '@nrwl/angular/testing/src/testing-utils';
+import { IOptions } from '../shared'; 
 import { classify } from '@angular-devkit/core/src/utils/strings';
 import * as pluralize from 'pluralize';
 
@@ -23,9 +22,7 @@ describe('imng-list', () => {
       storeName: 'employees',
       appPrefix: 'nrcrn',
     };
-    const tree: UnitTestTree = await readFirst(
-      runner.runSchematicAsync('imng-list', options, Tree.empty())
-    );
+    const tree: UnitTestTree = await runner.runSchematic('imng-list', options, Tree.empty());
 
     expect(tree.files).toEqual([
       `/test/${pluralize(options.name)}-list/index.ts`,
