@@ -6,8 +6,7 @@ import {
 import * as path from 'path';
 import { IOptions } from '../shared';
 import * as pluralize from 'pluralize';
-import { readFirst } from '@nrwl/angular/testing/src/testing-utils';
-import { classify, dasherize } from '@angular-devkit/core/src/utils/strings';
+import { classify, dasherize } from '@angular-devkit/core/src/utils/strings'; 
 
 const collectionPath = path.join(__dirname, `../collection.json`);
 
@@ -23,10 +22,9 @@ describe(`imng-module`, () => {
       swaggerProperties: [],
       appPrefix: 'aw',
     };
-    tree = await readFirst(
-      runner.runSchematicAsync(`imng-module`, options, Tree.empty())
-    );
-  })
+    tree = await runner.runSchematic(`imng-module`, options, Tree.empty());
+  });
+  
   test(`tree files should match`, () => {
     expect(tree.files).toMatchSnapshot();
   });
