@@ -5,8 +5,7 @@ import {
 } from '@angular-devkit/schematics/testing';
 import * as path from 'path';
 import { IOptions } from '../shared';
-import * as pluralize from 'pluralize';
-import { readFirst } from '@nrwl/angular/testing/src/testing-utils';
+import * as pluralize from 'pluralize'; 
 import { classify, dasherize } from '@angular-devkit/core/src/utils/strings';
 
 const collectionPath = path.join(__dirname, `../collection.json`);
@@ -23,9 +22,7 @@ describe(`imng-module`, () => {
       storeName: `employees`,
       appPrefix: '',
     };
-    const tree: UnitTestTree = await readFirst(
-      runner.runSchematicAsync(`imng-module`, options, Tree.empty())
-    );
+    const tree: UnitTestTree = await runner.runSchematic(`imng-module`, options, Tree.empty());
 
     expect(tree.files).toEqual([
       `/test/${pluralize(dasherize(options.name))}-module/${dasherize(
