@@ -4,8 +4,7 @@ import {
   UnitTestTree,
 } from '@angular-devkit/schematics/testing';
 import * as path from 'path';
-import { IOptions } from '../shared';
-import { readFirst } from '@nrwl/angular/testing/src/testing-utils';
+import { IOptions } from '../shared'; 
 import * as pluralize from 'pluralize';
 
 const collectionPath = path.join(__dirname, `../collection.json`);
@@ -19,9 +18,7 @@ describe(`imng-module`, () => {
       swaggerProperties: [],
       appPrefix: '',
     };
-    const tree: UnitTestTree = await readFirst(
-      runner.runSchematicAsync(`imng-module`, options, Tree.empty())
-    );
+    const tree: UnitTestTree = await runner.runSchematic(`imng-module`, options, Tree.empty());
 
     expect(tree.files).toEqual([
       `/test/${pluralize(options.name)}-module/${pluralize(
