@@ -5,7 +5,6 @@ import {
 } from '@angular-devkit/schematics/testing';
 import * as path from 'path';
 import { IOptions } from '../shared';
-import { readFirst } from '@nrwl/angular/testing/src/testing-utils';
 
 const collectionPath = path.join(__dirname, '../collection.json');
 
@@ -21,9 +20,7 @@ describe('imng-sub-list', () => {
       swaggerProperties: [],
       appPrefix: 'nrcrn',
     };
-    const tree: UnitTestTree = await readFirst(
-      runner.runSchematicAsync('imng-sub-list', options, Tree.empty())
-    );
+    const tree: UnitTestTree = await runner.runSchematic('imng-sub-list', options, Tree.empty());
 
     expect(tree.files).toEqual([
       `/test/+state/employee-certification.actions.ts`,
